Added auth checking to products and profiles

+++ products/views.py
@@ -1,11 +1,13 @@
 """ Describe this """
 from django.shortcuts import render, reverse, redirect, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
 from .models import Product, Category
 from .forms import ProductForm
+
 
 def all_products(request):
     """ View for all products, inc sorting & searching """
@@ -61,7 +63,6 @@ def all_products(request):
         "current_sorting": current_sorting,
     }
 
-
     return render(request, "products/products.html", context)
 
 
@@ -76,9 +77,13 @@ def product_detail(request, product_id):
 
     return render(request, "products/product_detail.html", context)
 
-
+@login_required
 def add_product(request):
     """ Add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, "Sorry, only store owners can do that")
+        return redirect(reverse("home)"))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -97,8 +102,14 @@ def add_product(request):
 
     return render(request, template, context)
 
+
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, "Sorry, only store owners can do that")
+        return redirect(reverse("home)"))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -120,12 +131,16 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
+
+@login_required
 def delete_product(request, product_id):
     """ Delere a product fromthe store """
+    if not request.user.is_superuser:
+        messages.error(request, "Sorry, only store owners can do that")
+        return redirect(reverse("home)"))
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
 
     messages.success(request, 'Successfully deleted product!')
 
-
-    return  redirect(reverse("products"))
+    return redirect(reverse("products"))
